docker/images/backend/requirements.txt
# ------------------- Web framework -------------------
fastapi                  # API principale (inclut Starlette)
uvicorn                  # serveur ASGI de prod / dev
python-dotenv            # chargement des variables d'environnement

# ------------------- Réseau / HTTP -------------------
httpx                    # client HTTP async utilisé dans le code
google-serp-api          # API pour interroger Google Search

# ------------------- Validation & configuration ------ 
pydantic                 # modèles de données
python-dotenv            # chargement des variables d'env (.env)

# ------------------- Uploads (multipart) -------------
python-multipart         # nécessaire à FastAPI pour File / Form

# ------------------- model -------------------------
torch                    # framework de deep learning
transformers             # bibliothèque pour les modèles de langage
accelerate               # accélération des modèles Transformers

pathlib                 # manipulation de chemins de fichiers
uuid                    # génération d'UUIDs
datetime                # manipulation de dates et heures